move internal logic to workdir pkg
- move WriteSpec to workdir
- init context
- use new function to init context
- reorder commands
- save spec
- move all logic to workdir
- change error
- fix spec search algorithm
- return an error when spec filename was not found
- change erros
